Add currency switching.

+++ cconverter/Models/IKCCurrency.h
@@ -7,7 +7,7 @@
 @property(nonatomic) NSString *fullName;
 @property(nonatomic) NSString *country;
 
-- (id)initWithIdentifier:(int)identifier
+- (id)initWithIdentifier:(NSInteger)identifier
             shortName:(NSString *)shortName
             fullName:(NSString *)fullName
             country:(NSString *)country;
